backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
torch
torchvision
transformers
pillow
numpy
